ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
tensorflow==2.15.0
# or pytorch if you prefer
# torch==2.1.1
# torchvision==0.16.1
python-dotenv==1.0.0
pydantic==2.5.0
